split out role auth from org auth; tidied up module dependencies
